Refactor async stream code. - Make everything plain extension methods. - Remove redundant methods.
